Move all parser tests to root parser/
