+++ src/test/java/seedu/address/logic/parser/DeleteTransactionCommandParserTest.java
@@ -1,4 +1,4 @@
-package seedu.address.logic.parser.transaction;
+package seedu.address.logic.parser;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
